refactor: separate hooks and default values
